Day 23 part 1 with long

+++ src/main/java/exercises/Day23.java
@@ -7,17 +7,17 @@ import java.util.List;
 
 public class Day23 {
     public static String es1(String rawInput, int moves) {
-        LinkedList<Integer> list = new LinkedList<>();
+        LinkedList<Long> list = new LinkedList<>();
         for (int i = 0; i < rawInput.length(); i++) {
-            list.addLast(Integer.parseInt(String.valueOf(rawInput.charAt(i))));
+            list.addLast(Long.parseLong(String.valueOf(rawInput.charAt(i))));
         }
 
-        int current = list.get(0);
+        long current = list.get(0);
         for (int move = 0; move < moves; move++) {
             int currentIndex = list.indexOf(current);
-            List<Integer> nextThree = getNextThree(list, currentIndex);
+            List<Long> nextThree = getNextThree(list, currentIndex);
 
-            int destination = current;
+            long destination = current;
             while (true) {
                 destination -= 1;
                 if (destination == 0) {
@@ -37,16 +37,16 @@ public class Day23 {
             current = list.get((list.indexOf(current) + 1) % list.size());
         }
 
-        int indexOfOne = list.indexOf(1);
-        List<Integer> firstHalf = list.subList(0, indexOfOne);
-        List<Integer> secondHalf = list.subList(indexOfOne + 1, list.size());
-        LinkedList<Integer> resultList = new LinkedList<>(secondHalf);
+        int indexOfOne = list.indexOf(1L);
+        List<Long> firstHalf = list.subList(0, indexOfOne);
+        List<Long> secondHalf = list.subList(indexOfOne + 1, list.size());
+        LinkedList<Long> resultList = new LinkedList<>(secondHalf);
         resultList.addAll(firstHalf);
         return StringUtils.join(resultList.toArray());
     }
 
-    private static List<Integer> getNextThree(LinkedList<Integer> list, int currentIndex) {
-        LinkedList<Integer> nextThree = new LinkedList<>();
+    private static List<Long> getNextThree(LinkedList<Long> list, int currentIndex) {
+        LinkedList<Long> nextThree = new LinkedList<>();
         for (int i = 1; i < 4; i++) {
             nextThree.add(list.get((currentIndex + i) % list.size()));
         }
